+++ FE/src/components/Layout.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect } from "react";
 
 import { BrowserRouter, Route } from "react-router-dom";
 
@@ -8,11 +8,28 @@ import Footer from "./Footer";
 import ProductViewModal from "./ProductViewModal";
 
 import Routes from "../routes/Routes";
+import { getUser } from "../api/product";
+import { useDispatch } from "react-redux";
+import { setUser } from "../redux/user-modal/userModalSlice";
 
 const Layout = () => {
-  sessionStorage.setItem("User", "HaiDuong");
-  const storgeValue = sessionStorage.getItem("username");
-  console.log(storgeValue);
+  const userName = sessionStorage.getItem("username");
+
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    const handleLogin = async () => {
+      if (userName) {
+        const userLogin = await getUser(userName);
+        dispatch(
+          setUser({ name: userLogin.tenDangNhap, role: userLogin.role })
+        );
+        console.log("userLogin: ", userLogin);
+      }
+    };
+    handleLogin();
+  }, []);
+
   return (
     <BrowserRouter>
       <Route
