+++ FE/src/pages/Admin.js
@@ -1,0 +1,9 @@
+const Admin = () => {
+  return (
+    <h1>
+      Admin Page
+    </h1>
+  );
+};
+
+export default Admin;
