+++ FE/src/pages/NotFoundPage.js
@@ -1,0 +1,7 @@
+const NotFoundPage = () => {
+    return (
+        <h1>NotFoundPage</h1>
+    )
+}
+
+export default NotFoundPage;
