+++ FE/src/pages/ProfileUser.js
@@ -1,0 +1,7 @@
+const ProfileUser = () => {
+    return (
+        <h1>ProfileUser</h1>
+    )
+}
+
+export default ProfileUser;
