+++ FE/src/redux/store.js
@@ -1,9 +1,12 @@
-import { configureStore } from '@reduxjs/toolkit'
-import productModalSlice from './product-modal/productModalSlice'
-import cartItemSlices from './shopping-cart/cartItemSlices'
+import { configureStore } from "@reduxjs/toolkit";
+import productModalSlice from "./product-modal/productModalSlice";
+import userModalSlice from "./user-modal/userModalSlice";
+import cartItemSlices from "./shopping-cart/cartItemSlices";
+
 export const store = configureStore({
-    reducer: {
-        productModal: productModalSlice,
-        cartItems: cartItemSlices
-    }
-})
+  reducer: {
+    productModal: productModalSlice,
+    cartItems: cartItemSlices,
+    user: userModalSlice,
+  },
+});
