+++ FE/src/routes/Routes.jsx
@@ -7,7 +7,14 @@ import Cart from "../pages/Cart";
 import Product from "../pages/Product";
 import Login from "../pages/Login";
 import Contact from "../pages/Contact";
+import Admin from "../pages/Admin";
+import ProfileUser from "../pages/ProfileUser";
+import NotFoundPage from "../pages/NotFoundPage";
+import { useSelector } from "react-redux";
+
 const Routes = () => {
+  const user = useSelector((state) => state.user);
+
   return (
     <Switch>
       <Route path="/" exact component={Home} />
@@ -16,6 +23,12 @@ const Routes = () => {
       <Route path="/cart" component={Cart} />
       <Route path="/login" component={Login} />
       <Route path="/contact" component={Contact} />
+      {user.role === 2 ? (
+        <Route path="/admin" component={Admin} />
+      ) : (
+        <Route path="/admin" component={NotFoundPage} />
+      )}
+      <Route path="/profile" component={ProfileUser} />
     </Switch>
   );
 };
